Communication style

This function detects the communication purpose and style of the text. The style includes:

self_revealing (sharing one's experience and opinion).

fact_oriented (focusing on factual information, objective observations or statements).

information_seeking (posing questions).

action_seeking (aiming to trigger someone's action by giving recommendation, requests or advice).

get_communic_style(url_api, text, language, all = FALSE, encoding = "UTF-8", token)

Arguments

url_api
URL to the API.

text
String with the texts. A set of maximum 32 texts is allowed.

language
Language of the text. Supported languages are "ar" (Arabic), "de" (German), "en" (English), "es" (Spanish), "fr" (French), "it" (Italian), "nl" (Dutch), "pt" (Portuguese), "ru" (Russian), "tr" (Turkish) and "zh" (Chinese).

all
Boolean. If TRUE, all the probabilities are displayed. If FALSE, only the largest probability is displayed.

encoding
Character encoding. Default "UTF-8".

token
The access token.

Value

A data frame with the texts and the probabilities of the psychological features obtained.

Note

The type of array for one text is:
data <- "[\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"]"
For more than one text is:
data <- "[\"id\":\"1\",\"text\":\"love\",\"language\":\"en\",
\"id\":\"2\",\"text\":\"hate\",\"language\":\"en\"]"
(write it in one line). In the first case, add a curly bracket after the first box bracket and before the last box bracket. In the second case, add the curly brackets for every set of id, text and language.

Author

Symanto

Examples

if (FALSE) {
token <- "Request your API Key: https://developers.symanto.net/signup"
url_api <- "https://api.symanto.net"

# For only one text:
# English:
text <- "I love the service."
language <- "en"
get_communic_style(url_api, text, language, all = TRUE, "UTF-8", token)
get_communic_style(url_api, text, language, all = FALSE, "UTF-8", token)

# German:
text <- "Menschen langweilen sich zu Hause."
language <- "de"
get_communic_style(url_api, text, language, TRUE, "UTF-8", token)
get_communic_style(url_api, text, language, FALSE, "UTF-8", token)

# Spanish:
text <- "Estoy muy contento."
language <- "es"
get_communic_style(url_api, text, language, TRUE, "UTF-8", token)
get_communic_style(url_api, text, language, FALSE, "UTF-8", token)

# For more than one text:
# English:
df <- data.frame(text = c("love", "hate"), language = c("en", "en"))
text <- df$text
language <- df$language
get_communic_style(url_api, text, language, all = TRUE, "UTF-8", token)
get_communic_style(url_api, text, language, all = FALSE, "UTF-8", token)
}