Label texts

Given a certain (Excel) file with some posts, this function assigns the psychological features that the user wants to. In the interests of computational speed, the whole set of texts is splitted into smaller blocks.

label_text(url_api, excel_file, user_actions, encoding = "UTF-8", token, 
           texts_in_blocks, user_language = "en", verbose = FALSE)

Arguments

url_api
URL to the API.

excel_file
Excel file. It must contain at least one column with the texts, named "Text" or "text". Another column with the language of the texts, named "Language" or "language", is optional (see the user_language parameter below). Other complementary columns can be also added to be incorporated in the returned data frame.

user_actions
Vector with the set of psychological properties that the user wants to obtain. Possible values are "personality", "communication", "emotions", "ekman_emotions", "sentiments", "topic_sentiments" and "language".

encoding
Character encoding. Default "UTF-8".

token
The access token.

texts_in_blocks
Number of texts in every block. Basically, if there are 50 texts and texts_in_blocks is 25, there will be two blocks with 25 texts. The procedure will be applied to both of them iteratively. This method is much faster than going text by text.

user_language
If the language of the texts is not available, get_language_detection will be used to assign a language to every text. For the few cases where this tool does not return a valid language, the assigned language is this parameter provided by the user. In the case when the language is available, this parameter serves to replace those languages that are not allowed by the API. Default "en" (English).

verbose
Provides a flow of information.

Value

The excel_file filled with the probablities related to the psychological properties desired by the user.

Details

The personality traits are emotional and rational.

The communication style are action_seeking, fact_oriented, information_seeking and self_revealing.

The emotions are anger, joy, love, sadness and surprise.

The Ekman's emotions are anger, disgust, fear, joy and no_emotion, sadness and surprise.

The sentiments are positive and negative.

Note

The accepted codes in the Symanto API for user_language are as follows: "ar" (Arabic), "de" (German), "en" (English), "es" (Spanish), "fr" (French), "it" (Italian), "nl" (Dutch), "pt" (Portuguese), "ru" (Russian), "tr" (Turkish) and "zh" (Chinese).

The lowest common denominator for creating blocks of texts in every function is 32, so this should be the largest possible number in texts_in_blocks when all the psychological properties are selected in user_actions. However, the Ekman's emotions seem not to accept this number of texts, so we recommend to choose 5 if the Ekman's emotions belong to the desired user_actions.

Due to the internal procedure of the computational model, only blocks of texts of the same language are created.

See also

Author

Symanto

Examples

if (FALSE) {
#library(openxlsx)
token <- "Request your API Key: https://developers.symanto.net/signup"
url_api <- "https://api.symanto.net"

# All actions:
user_actions <- c("personality", "communication", "emotions", "ekman_emotions", 
                  "sentiments", "topic_sentiments", "language")  
excel_dest_l1 <- label_text(url_api, excel_orig, user_actions, "UTF-8", 
                            token, 5, "en", TRUE)
#write.xlsx(excel_dest_l1, "excel_dest_l1.xlsx")

# A sample of actions:
user_actions <- c("personality", "sentiments")  
excel_dest_l2 <- label_text(url_api, excel_orig, user_actions, "UTF-8", 
                            token, 32, "en", TRUE)
#write.xlsx(excel_dest_l2, "excel_dest_l2.xlsx")

excel_orig1 <- excel_orig[, c("id", "text")]
user_actions <- c("personality", "sentiments", "language")  
excel_dest_l3 <- label_text(url_api, excel_orig1, user_actions, "UTF-8", 
                            token, 32, "en", TRUE)
#write.xlsx(excel_dest_l3, "excel_dest_l3.xlsx")

# Case when one the input languages is not valid:
user_actions <- c("personality", "sentiments")  
excel_orig2 <- excel_orig[1:3,]
excel_orig2$language[1] <- "es-1"
excel_dest_l4 <- label_text(url_api, excel_orig2, user_actions, "UTF-8", 
                            token, 32, "en", TRUE)
#write.xlsx(excel_dest_l4, "excel_dest_l4.xlsx")
}